Using mkDerivation to build Linux keymaps, directly from GitHub!
I created a project on GitHub to generate Linux console keymaps. I
wanted to use it in my NixOS configuration.nix, so I wrote this
default.nix to test using nix-build.

Once I got it working, I basically copied the contents of default.nix
into the beginning of my configuration.nix. It worked great!

If you want to see this in action, just install Nix and nixpkgs and do
this:

   nix-build